+++ internal/infrastructure/repositories_sqlite.go
@@ -19,14 +19,14 @@ func NewRepositoriesSQLite(db *sql.DB) *handlers.Repositories {
 	userRepo := user.NewSQLiteRepository(db)
 	categoryRepo := category.NewSQLiteRepository(db)
 	return &handlers.Repositories{
-		User:           userRepo,
-		Family:         user.NewSQLiteFamilyRepository(db, categoryRepo, userRepo),
-		Category:       categoryRepo,
-		Account:        account.NewSQLiteRepository(db),
-		Holding:        holding.NewSQLiteRepository(db),
-		Reconciliation: reconciliation.NewSQLiteRepository(db),
-		Transaction:    transaction.NewSQLiteRepository(db),
-		Budget:         budget.NewSQLiteRepository(db),
-		Session:        auth.NewSessionSQLiteRepository(db),
+		User:        userRepo,
+		Family:      user.NewSQLiteFamilyRepository(db, categoryRepo, userRepo),
+		Category:    categoryRepo,
+		Account:     account.NewSQLiteRepository(db),
+		Holding:     holding.NewSQLiteRepository(db),
+		Balance:     reconciliation.NewBalanceSQLiteRepository(db),
+		Transaction: transaction.NewSQLiteRepository(db),
+		Budget:      budget.NewSQLiteRepository(db),
+		Session:     auth.NewSessionSQLiteRepository(db),
 	}
 }